requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
PyYAML>=6.0
numpy>=1.24
matplotlib>=3.8
pandas>=2.0
